chore(aggregated-assets): delete the dada-client shims (LIVE-35231)
Every consumer points at the new packages, so the 26-file shim tree goes. Until
now there were two valid import paths for the same code.

The tree still held the only coverage for eight symbols, so three test files
moved into @domain/api-aggregated-assets rather than being deleted with it.
Nothing there duplicated the domain package's own tests: it had no coverage at
all for buildAssetsQueryParams, the two category accessors, convertApiAssets,
transformAssetsResponse, or the five error helpers - 74 tests, including the
leniency behaviour this epic has been careful to preserve. git reports both
moved files as 99% renames, so only their import paths changed. The third was
merged into the existing api.test.ts because its fixtures are shared across all
three describe blocks and splitting them would have meant duplicating forty
lines. domain/api goes from 84 tests to 158.

Only one of the four dependencies the ticket lists could actually be dropped.
LIVE-35230 pointed modularDrawer and portfolio directly at
@domain/api-aggregated-assets, @domain/entity-interest-rate and
@features/platform-aggregated-assets, so live-common still needs all three;
only @domain/entity-aggregated-asset is now unreferenced. Removing it also
retires the knip ignoreUnused entry added in LIVE-35230, along with 24 stale
ignoreUnimported paths pointing into the deleted tree.

The DADA service description needed no rescuing: domain/api's README already
defines the acronym and the six-collection payload.
