+++ .changeset/lucky-foxes-delete.md
@@ -1,0 +1,6 @@
+---
+"@ledgerhq/live-common": patch
+"@domain/api-aggregated-assets": patch
+---
+
+Delete the dada-client shim tree; the DADA code now lives only in its DDD packages
