+++ domain/api/aggregated-assets/src/api.test.ts
@@ -7,7 +7,15 @@ jest.mock("@shared/env", () => ({
   getEnv: jest.fn().mockReturnValue("https://dada.api.ledger.com/v1"),
 }));
 
-import { assetsDataApi } from "./api";
+import { getEnv } from "@shared/env";
+import {
+  assetsDataApi,
+  buildAssetsQueryParams,
+  fetchAllAssetsByCategory,
+  fetchAllAssetCurrencyIdsByCategory,
+} from "./index";
+import { AssetCategory } from "./types";
+import type { RawApiResponse } from "./schema";
 
 describe("assetsDataApi", () => {
   /*
@@ -17,5 +25,245 @@ describe("assetsDataApi", () => {
    */
   it("keeps the frozen reducerPath", () => {
     expect(assetsDataApi.reducerPath).toBe("assetsDataApi");
+  });
+});
+
+const getEnvMock = jest.mocked(getEnv);
+
+function makePage(tickers: string[]): RawApiResponse {
+  const cryptoAssets: RawApiResponse["cryptoAssets"] = {};
+  for (const ticker of tickers) {
+    cryptoAssets[ticker.toLowerCase()] = {
+      id: ticker.toLowerCase(),
+      ticker,
+      name: ticker,
+      assetsIds: {},
+    };
+  }
+  return {
+    cryptoAssets,
+    networks: {},
+    cryptoOrTokenCurrencies: {},
+    interestRates: {},
+    markets: {},
+    currenciesOrder: { key: "marketCap", order: "desc", metaCurrencyIds: [] },
+  };
+}
+
+function okResponse(body: RawApiResponse, nextCursor?: string): Response {
+  const headers = new Headers();
+  if (nextCursor) headers.set("x-ledger-next", nextCursor);
+  return new Response(JSON.stringify(body), { status: 200, headers });
+}
+
+function errorResponse(status: number, statusText: string): Response {
+  return new Response(null, { status, statusText });
+}
+
+function mockFetchPages(pages: { tickers: string[]; nextCursor?: string }[]): jest.SpyInstance {
+  const spy = jest.spyOn(globalThis, "fetch");
+  for (const page of pages) {
+    spy.mockResolvedValueOnce(okResponse(makePage(page.tickers), page.nextCursor));
+  }
+  return spy;
+}
+
+const defaultArgs = {
+  category: AssetCategory.Stablecoins,
+  product: "lld" as const,
+  version: "1.0.0",
+};
+
+const baseQueryArg = { product: "lld" as const, version: "1.0.0" };
+
+describe("buildAssetsQueryParams", () => {
+  it("should pass currencyIds through unchanged", () => {
+    const input = [
+      "ethereum/erc20/usdc",
+      "stellar/asset/USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN",
+      "multiversx/esdt/USDC-c76f1f",
+      "ton/jetton/some_address",
+      "bitcoin",
+    ];
+    const params = buildAssetsQueryParams({ ...baseQueryArg, currencyIds: input });
+    expect(params.currencyIds).toEqual(input);
+  });
+
+  it("should omit currencyIds when empty or not provided", () => {
+    expect(
+      buildAssetsQueryParams({ ...baseQueryArg, currencyIds: [] }).currencyIds,
+    ).toBeUndefined();
+    expect(buildAssetsQueryParams(baseQueryArg).currencyIds).toBeUndefined();
+  });
+
+  it("should serialize networkIds as a comma-separated string", () => {
+    const input = ["ethereum", "tron"];
+
+    expect(buildAssetsQueryParams({ ...baseQueryArg, networkIds: input }).networkIds).toBe(
+      "ethereum,tron",
+    );
+  });
+
+  it("should omit networkIds when empty or not provided", () => {
+    expect(buildAssetsQueryParams({ ...baseQueryArg, networkIds: [] }).networkIds).toBeUndefined();
+    expect(buildAssetsQueryParams(baseQueryArg).networkIds).toBeUndefined();
+  });
+
+  it("should serialize categories as a comma-separated string", () => {
+    expect(
+      buildAssetsQueryParams({ ...baseQueryArg, categories: [AssetCategory.Stocks] }).categories,
+    ).toBe("stocks");
+    expect(
+      buildAssetsQueryParams({
+        ...baseQueryArg,
+        categories: [AssetCategory.Stocks, AssetCategory.Stablecoins],
+      }).categories,
+    ).toBe("stocks,stablecoins");
+  });
+
+  it("should omit categories when empty or not provided", () => {
+    expect(buildAssetsQueryParams({ ...baseQueryArg, categories: [] }).categories).toBeUndefined();
+    expect(buildAssetsQueryParams(baseQueryArg).categories).toBeUndefined();
+  });
+});
+
+describe("fetchAllAssetsByCategory", () => {
+  afterEach(() => {
+    jest.restoreAllMocks();
+  });
+
+  it("should aggregate tickers from 3 pages using x-ledger-next cursor", async () => {
+    const spy = mockFetchPages([
+      { tickers: ["USDT", "USDC"], nextCursor: "cursor-2" },
+      { tickers: ["DAI", "BUSD"], nextCursor: "cursor-3" },
+      { tickers: ["TUSD", "FRAX"] },
+    ]);
+
+    const result = await fetchAllAssetsByCategory(defaultArgs);
+
+    expect(result).toEqual({ data: ["USDT", "USDC", "DAI", "BUSD", "TUSD", "FRAX"] });
+    expect(spy).toHaveBeenCalledTimes(3);
+
+    expect(new URL(String(spy.mock.calls[0][0])).searchParams.has("cursor")).toBe(false);
+    expect(new URL(String(spy.mock.calls[1][0])).searchParams.get("cursor")).toBe("cursor-2");
+    expect(new URL(String(spy.mock.calls[2][0])).searchParams.get("cursor")).toBe("cursor-3");
+  });
+
+  it("should return error and stop when a page fails", async () => {
+    const spy = jest.spyOn(globalThis, "fetch");
+    spy.mockResolvedValueOnce(okResponse(makePage(["USDT"]), "cursor-2"));
+    spy.mockResolvedValueOnce(errorResponse(502, "Bad Gateway"));
+
+    const result = await fetchAllAssetsByCategory(defaultArgs);
+
+    expect(result).toEqual({
+      error: { status: 502, data: "Failed to fetch assets by category: Bad Gateway" },
+    });
+    expect(spy).toHaveBeenCalledTimes(2);
+  });
+
+  it("should return FETCH_ERROR on network failure", async () => {
+    jest.spyOn(globalThis, "fetch").mockRejectedValueOnce(new Error("Network failure"));
+
+    const result = await fetchAllAssetsByCategory(defaultArgs);
+
+    expect(result).toEqual({
+      error: { status: "FETCH_ERROR", error: "Network failure" },
+    });
+  });
+
+  it("should block requests to untrusted hosts", async () => {
+    getEnvMock.mockImplementation(() => "https://evil.example.com/v1");
+    const spy = jest.spyOn(globalThis, "fetch");
+
+    const result = await fetchAllAssetsByCategory(defaultArgs);
+
+    expect(result).toEqual({
+      error: {
+        status: "FETCH_ERROR",
+        error: "Blocked request to untrusted host: evil.example.com",
+      },
+    });
+    expect(spy).not.toHaveBeenCalled();
+  });
+});
+
+function makePageWithIds(
+  assets: { key: string; assetsIds: Record<string, string> }[],
+): RawApiResponse {
+  const cryptoAssets: RawApiResponse["cryptoAssets"] = {};
+  for (const { key, assetsIds } of assets) {
+    cryptoAssets[key] = { id: key, ticker: key.toUpperCase(), name: key, assetsIds };
+  }
+  return {
+    cryptoAssets,
+    networks: {},
+    cryptoOrTokenCurrencies: {},
+    interestRates: {},
+    markets: {},
+    currenciesOrder: { key: "marketCap", order: "desc", metaCurrencyIds: [] },
+  };
+}
+
+const stocksArgs = { ...defaultArgs, category: AssetCategory.Stocks };
+
+describe("fetchAllAssetCurrencyIdsByCategory", () => {
+  beforeEach(() => {
+    getEnvMock.mockReturnValue("https://dada.api.ledger.com/v1");
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
+  });
+
+  it("should aggregate and flatten currency ids across pages", async () => {
+    const spy = jest.spyOn(globalThis, "fetch");
+    spy.mockResolvedValueOnce(
+      okResponse(
+        makePageWithIds([{ key: "aapl", assetsIds: { ethereum: "eth/aapl", solana: "sol/aapl" } }]),
+        "cursor-2",
+      ),
+    );
+    spy.mockResolvedValueOnce(
+      okResponse(makePageWithIds([{ key: "tsla", assetsIds: { ethereum: "eth/tsla" } }])),
+    );
+
+    const result = await fetchAllAssetCurrencyIdsByCategory(stocksArgs);
+
+    expect(result).toEqual({ data: ["eth/aapl", "sol/aapl", "eth/tsla"] });
+    expect(spy).toHaveBeenCalledTimes(2);
+  });
+
+  it("should skip assets that have no assetsIds", async () => {
+    jest.spyOn(globalThis, "fetch").mockResolvedValueOnce(
+      okResponse(
+        makePageWithIds([
+          { key: "aapl", assetsIds: {} },
+          { key: "tsla", assetsIds: { ethereum: "eth/tsla" } },
+        ]),
+      ),
+    );
+
+    const result = await fetchAllAssetCurrencyIdsByCategory(stocksArgs);
+
+    expect(result).toEqual({ data: ["eth/tsla"] });
+  });
+
+  it("should return error and stop when a page fails", async () => {
+    const spy = jest.spyOn(globalThis, "fetch");
+    spy.mockResolvedValueOnce(
+      okResponse(
+        makePageWithIds([{ key: "aapl", assetsIds: { ethereum: "eth/aapl" } }]),
+        "cursor-2",
+      ),
+    );
+    spy.mockResolvedValueOnce(errorResponse(502, "Bad Gateway"));
+
+    const result = await fetchAllAssetCurrencyIdsByCategory(stocksArgs);
+
+    expect(result).toEqual({
+      error: { status: 502, data: "Failed to fetch assets by category: Bad Gateway" },
+    });
+    expect(spy).toHaveBeenCalledTimes(2);
   });
 });
