+++ libs/ledger-live-common/.unimportedrc.json
@@ -416,30 +416,6 @@
     "src/config/sharedConfig.ts",
     "src/counterValues/hooks/useUsdToFiatRate.ts",
     "src/currencies/hooks.ts",
-    "src/dada-client/entities/index.ts",
-    "src/dada-client/entities/interestRateSelectors.ts",
-    "src/dada-client/entities/marketSelectors.ts",
-    "src/dada-client/entities/selectorUtils.ts",
-    "src/dada-client/hooks/useAssetData.ts",
-    "src/dada-client/hooks/useAssetsData.ts",
-    "src/dada-client/hooks/useChunkedAssetsData.ts",
-    "src/dada-client/hooks/useDrawerConfiguration.ts",
-    "src/dada-client/hooks/useInterestRatesByCurrencies.ts",
-    "src/dada-client/hooks/useLazyLedgerCurrency.ts",
-    "src/dada-client/hooks/useMarketByCurrencies.ts",
-    "src/dada-client/hooks/useStablecoinTickers.ts",
-    "src/dada-client/hooks/useStockAssetIds.ts",
-    "src/dada-client/hooks/useStocksData.ts",
-    "src/dada-client/index.ts",
-    "src/dada-client/state-manager/api.ts",
-    "src/dada-client/state-manager/types.ts",
-    "src/dada-client/types/trend.ts",
-    "src/dada-client/utils/assetDiscovery.ts",
-    "src/dada-client/utils/chunkCurrencyIds.ts",
-    "src/dada-client/utils/currencySelection.ts",
-    "src/dada-client/utils/deepMergeCryptoAssets.ts",
-    "src/dada-client/utils/errorUtils.ts",
-    "src/dada-client/utils/mergeAssetsDataPages.ts",
     "src/device",
     "src/device-action/utils.ts",
     "src/device/hooks/useLatestFirmware.ts",
@@ -831,7 +807,6 @@
   ],
   "ignoreUnused": [
     "@cardano-foundation/ledgerjs-hw-app-cardano",
-    "@domain/entity-aggregated-asset",
     "@ledgerhq/asset-aggregation",
     "@ledgerhq/coin-hypercore",
     "@ledgerhq/coin-zcash",
