+++ libs/ledger-live-common/package.json
@@ -266,7 +266,7 @@
     "@domain/api-aggregated-assets": "workspace:^",
     "@domain/api-currency-token": "workspace:^",
     "@domain/api-swap-quotes": "workspace:^",
-    "@domain/entity-aggregated-asset": "workspace:^",
+    "@domain/entity-account-name": "workspace:^",
     "@domain/entity-client-identity": "workspace:^",
     "@domain/entity-currency": "workspace:^",
     "@domain/entity-currency-crypto": "workspace:^",
@@ -274,6 +274,7 @@
     "@domain/entity-currency-token": "workspace:^",
     "@domain/entity-currency-unit": "workspace:^",
     "@domain/entity-interest-rate": "workspace:^",
+    "@domain/entity-recent-addresses": "workspace:^",
     "@features/platform-aggregated-assets": "workspace:*",
     "@features/platform-env": "workspace:^",
     "@features/platform-feature-flags": "workspace:^",
@@ -318,8 +319,6 @@
     "@ledgerhq/devices": "workspace:*",
     "@ledgerhq/domain-service": "workspace:^",
     "@ledgerhq/evm-tools": "workspace:^",
-    "@domain/entity-account-name": "workspace:^",
-    "@domain/entity-recent-addresses": "workspace:^",
     "@ledgerhq/hw-app-algorand": "workspace:^",
     "@ledgerhq/hw-app-aptos": "workspace:^",
     "@ledgerhq/hw-app-btc": "workspace:^",
