--- libs/ledger-live-common/src/dada-client/__mocks__/assets.mock.ts
@@ -1,1 +1,0 @@
-export * from "@domain/api-aggregated-assets/mock";
