--- libs/ledger-live-common/src/dada-client/entities/index.ts
@@ -1,11 +1,0 @@
-export type { CryptoAssetMeta } from "@domain/entity-aggregated-asset";
-export type { InterestRate } from "@domain/entity-interest-rate";
-export type {
-  ApiAsset,
-  ApiCryptoCurrency,
-  ApiTokenCurrency,
-  AssetsData,
-  CurrenciesOrder,
-  NetworkInfo,
-  RawApiResponse,
-} from "@domain/api-aggregated-assets";
