--- libs/ledger-live-common/src/dada-client/hooks/useAssetData.ts
@@ -1,1 +1,0 @@
-export { useAssetData } from "@features/platform-aggregated-assets";
