--- libs/ledger-live-common/src/dada-client/hooks/useAssetsData.ts
@@ -1,1 +1,0 @@
-export { useAssetsData } from "@features/platform-aggregated-assets";
