--- libs/ledger-live-common/src/dada-client/hooks/useChunkedAssetsData.ts
@@ -1,1 +1,0 @@
-export { useChunkedAssetsData } from "@features/platform-aggregated-assets";
