--- libs/ledger-live-common/src/dada-client/hooks/useDrawerConfiguration.ts
@@ -1,1 +1,0 @@
-export { useDrawerConfiguration } from "../../wallet-api/ModularDrawer/useDrawerConfiguration";
