--- libs/ledger-live-common/src/dada-client/hooks/useInterestRatesByCurrencies.ts
@@ -1,2 +1,0 @@
-export { useInterestRatesByCurrencies } from "@features/platform-aggregated-assets";
-export type { InterestRatesByCurrencies } from "@features/platform-aggregated-assets";
