--- libs/ledger-live-common/src/dada-client/hooks/useLazyLedgerCurrency.ts
@@ -1,1 +1,0 @@
-export { useLazyLedgerCurrency } from "@features/platform-aggregated-assets";
