--- libs/ledger-live-common/src/dada-client/hooks/useMarketByCurrencies.ts
@@ -1,1 +1,0 @@
-export { useMarketByCurrencies } from "@features/platform-aggregated-assets";
