--- libs/ledger-live-common/src/dada-client/hooks/useStablecoinTickers.ts
@@ -1,1 +1,0 @@
-export { useStablecoinTickers } from "@features/platform-aggregated-assets";
