--- libs/ledger-live-common/src/dada-client/hooks/useStockAssetIds.ts
@@ -1,1 +1,0 @@
-export { useStockAssetIds } from "@features/platform-aggregated-assets";
