--- libs/ledger-live-common/src/dada-client/hooks/useStocksData.ts
@@ -1,1 +1,0 @@
-export { useStocksData } from "@features/platform-aggregated-assets";
