--- libs/ledger-live-common/src/dada-client/index.ts
@@ -1,1 +1,0 @@
-export * from "./entities";
