--- libs/ledger-live-common/src/dada-client/mocks/stablecoins.mock.ts
@@ -1,1 +1,0 @@
-export * from "@domain/api-aggregated-assets/mock/stablecoins";
