--- libs/ledger-live-common/src/dada-client/state-manager/api.ts
@@ -1,1 +1,0 @@
-export * from "@domain/api-aggregated-assets";
