--- libs/ledger-live-common/src/dada-client/state-manager/types.ts
@@ -1,13 +1,0 @@
-export {
-  AssetCategory,
-  AssetsAdditionalData,
-  AssetsDataTags,
-  ONE_DAY_IN_SECONDS,
-} from "@domain/api-aggregated-assets";
-export type {
-  AssetsData,
-  AssetsDataWithPagination,
-  GetAssetsByCategoryParams,
-  GetAssetsDataParams,
-  PageParam,
-} from "@domain/api-aggregated-assets";
