--- libs/ledger-live-common/src/dada-client/types/trend.ts
@@ -1,1 +1,0 @@
-export type { ApyType } from "@domain/entity-interest-rate";
