--- libs/ledger-live-common/src/dada-client/utils/assetDiscovery.ts
@@ -1,6 +1,0 @@
-export { selectTopStocks, selectTopAssetsByCategory } from "@features/platform-aggregated-assets";
-export type {
-  StockSuggestion,
-  CategorizedDiscoveryAsset,
-  CategorizedDiscoveryAssets,
-} from "@features/platform-aggregated-assets";
