--- libs/ledger-live-common/src/dada-client/utils/currencySelection.ts
@@ -1,1 +1,0 @@
-export { selectCurrency, selectCurrencyForMetaId } from "@features/platform-aggregated-assets";
