--- libs/ledger-live-common/src/dada-client/utils/errorUtils.ts
@@ -1,8 +1,0 @@
-export {
-  getApiErrorStatus,
-  isApiError,
-  isFetchBaseQueryError,
-  isNetworkError,
-  parseError,
-} from "@domain/api-aggregated-assets";
-export type { ErrorInfo } from "@domain/api-aggregated-assets";
