--- libs/ledger-live-common/src/dada-client/utils/mergeAssetsDataPages.ts
@@ -1,1 +1,0 @@
-export { mergeAssetsDataPages } from "@domain/api-aggregated-assets";
